requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn[standard]
gunicorn

# Core Dependencies
python-dotenv
requests
httpx

# Data Handling
pandas
simplejson

# Database (if needed for future use)
sqlalchemy
alembic
psycopg2-binary==2.9.9
psycopg[pool]==3.2.6   # for connection pooling
asyncpg==0.31.0
requests

# LLM
openai
pydantic==2.12.5
tiktoken
tenacity
transformers
# Logfire
logfire

# Caching
redis
aiocache

# Authentication (FastAPI compatible)
PyJWT
cryptography

# Language and Localization
langcodes
language-data

# File Processing
openpyxl

# Template Engine
jinja2

# Pydantic AI 
pydantic-ai==1.42.0
google-genai==1.57.0
asgiref


gunicorn
uvicorn
httpx
marqo==3.11.0

# Tenacity
tenacity
langcodes

# AWS
boto3

# Azure
azure-cognitiveservices-speech==1.47.0

# VAD
soundfile
torch==2.9.1
torchaudio==2.9.1

# Mapbox Geocoding
geopy

rapidfuzz
fastapi
aiocache

# Pipecat
pipecat-ai[silero,rnnoise]
onnxruntime

# data conversion
convertdate

# Cosdata Vector Database Client
cosdata-client>=0.2.2

# Sentence Transformers for local embeddings (Cosdata RAG)
sentence-transformers>=2.2.0

# Language data for langcodes
language-data>=1.4.0
